+++ Gen6/Dumper6.cs
@@ -110,7 +110,7 @@ public static class Dumper6
         using var ms = new MemoryStream();
         using var bw = new BinaryWriter(ms);
 
-        bw.Write((ushort)area.Location);
+        bw.Write(area.Location);
         bw.Write((byte)area.Type);
         bw.Write((byte)0);
 
@@ -123,8 +123,8 @@ public static class Dumper6
     private static void WriteSlot(BinaryWriter bw, EncounterSlot6 slot)
     {
         bw.Write((ushort)(slot.Species | (slot.Form << 11)));
-        bw.Write((byte)slot.LevelMin);
-        bw.Write((byte)slot.LevelMax);
+        bw.Write(slot.LevelMin);
+        bw.Write(slot.LevelMax);
     }
 
     #region XY Alt Slots
