+++ Gen7/Dumper7.cs
@@ -87,7 +87,7 @@ public static class Dumper7
         using var ms = new MemoryStream();
         using var bw = new BinaryWriter(ms);
 
-        bw.Write((ushort)area.Location);
+        bw.Write(area.Location);
         bw.Write((byte)area.Type);
         bw.Write((byte)0);
 
@@ -100,8 +100,8 @@ public static class Dumper7
     private static void WriteSlot(BinaryWriter bw, EncounterSlot7 slot)
     {
         bw.Write((ushort)(slot.Species | (slot.Form << 11)));
-        bw.Write((byte)slot.LevelMin);
-        bw.Write((byte)slot.LevelMax);
+        bw.Write(slot.LevelMin);
+        bw.Write(slot.LevelMax);
     }
 
     private static void MarkAreasAsSOS(ref EncounterArea7[] Areas)
